restart the branch after a discussion with @RiccardoM on cyclic dependency between reactions and post modules
