+++ x/posts/internal/keeper/keeper.go
@@ -1,13 +1,12 @@
 package keeper
 
 import (
-	"bytes"
-	"fmt"
 	"sort"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/desmos-labs/desmos/x/posts/internal/types"
 )
 
@@ -183,160 +182,4 @@ func (k Keeper) GetPostsFiltered(ctx sdk.Context, params types.QueryPostsParams)
 	}
 
 	return filteredPosts
-}
-
-// SavePollAnswers save the poll's answers associated with the given postID inside the current context
-// It assumes that the post exists and has a Poll inside it.
-// If userAnswersDetails are already present, the old ones will be overridden.
-func (k Keeper) SavePollAnswers(ctx sdk.Context, postID types.PostID, userPollAnswers types.UserAnswer) {
-	store := ctx.KVStore(k.StoreKey)
-
-	sort.Slice(
-		userPollAnswers.Answers,
-		func(i, j int) bool { return userPollAnswers.Answers[i] < userPollAnswers.Answers[j] },
-	)
-
-	usersAnswersDetails := k.GetPollAnswers(ctx, postID)
-
-	if usersAnswersDetails, appended := usersAnswersDetails.AppendIfMissingOrIfUsersEquals(userPollAnswers); appended {
-		store.Set(types.PollAnswersStoreKey(postID), k.Cdc.MustMarshalBinaryBare(&usersAnswersDetails))
-	}
-
-}
-
-// GetPollAnswers returns the list of all the post polls answers associated with the given postID that are stored into the current state.
-func (k Keeper) GetPollAnswers(ctx sdk.Context, postID types.PostID) types.UserAnswers {
-	store := ctx.KVStore(k.StoreKey)
-
-	var usersAnswersDetails types.UserAnswers
-	answersBz := store.Get(types.PollAnswersStoreKey(postID))
-
-	k.Cdc.MustUnmarshalBinaryBare(answersBz, &usersAnswersDetails)
-
-	return usersAnswersDetails
-}
-
-// GetPollAnswersMap allows to returns the list of answers that have been stored inside the given context
-func (k Keeper) GetPollAnswersMap(ctx sdk.Context) map[types.PostID]types.UserAnswers {
-	store := ctx.KVStore(k.StoreKey)
-	iterator := sdk.KVStorePrefixIterator(store, types.PollAnswersStorePrefix)
-
-	usersAnswersData := map[types.PostID]types.UserAnswers{}
-	for ; iterator.Valid(); iterator.Next() {
-		var userAnswers types.UserAnswers
-		k.Cdc.MustUnmarshalBinaryBare(iterator.Value(), &userAnswers)
-		idBytes := bytes.TrimPrefix(iterator.Key(), types.PollAnswersStorePrefix)
-		postID, err := types.ParsePostID(string(idBytes))
-		if err != nil {
-			// This should never happen
-			panic(err)
-		}
-
-		usersAnswersData[postID] = userAnswers
-	}
-
-	return usersAnswersData
-}
-
-// GetPollAnswersByUser retrieves post poll answers associated to the given ID and filtered by user
-func (k Keeper) GetPollAnswersByUser(ctx sdk.Context, postID types.PostID, user sdk.AccAddress) []types.AnswerID {
-	postPollAnswers := k.GetPollAnswers(ctx, postID)
-
-	for _, postPollAnswers := range postPollAnswers {
-		if user.Equals(postPollAnswers.User) {
-			return postPollAnswers.Answers
-		}
-	}
-	return nil
-}
-
-// -------------
-// --- Reactions
-// -------------
-
-// SaveReaction allows to save the given reaction inside the store.
-// It assumes that the given reaction is valid.
-// If another reaction from the same user for the same post and with the same value exists, returns an expError.
-// nolint: interfacer
-func (k Keeper) SaveReaction(ctx sdk.Context, postID types.PostID, reaction types.Reaction) error {
-	store := ctx.KVStore(k.StoreKey)
-	key := types.PostReactionsStoreKey(postID)
-
-	// Get the existent reactions
-	var reactions types.Reactions
-	k.Cdc.MustUnmarshalBinaryBare(store.Get(key), &reactions)
-
-	// Check for double reactions
-	if reactions.ContainsReactionFrom(reaction.Owner, reaction.Value) {
-		return fmt.Errorf("%s has already reacted with %s to the post with id %s",
-			reaction.Owner, reaction.Value, postID)
-	}
-
-	// Save the new reaction
-	reactions = append(reactions, reaction)
-	store.Set(key, k.Cdc.MustMarshalBinaryBare(&reactions))
-
-	return nil
-}
-
-// RemoveReaction removes the reaction from the given user from the post having the
-// given postID. If no reaction with the same value was previously added from the given user, an expError
-// is returned.
-// nolint: interfacer
-func (k Keeper) RemoveReaction(ctx sdk.Context, postID types.PostID, user sdk.AccAddress, value string) error {
-	store := ctx.KVStore(k.StoreKey)
-	key := types.PostReactionsStoreKey(postID)
-
-	// Get the existing reactions
-	var reactions types.Reactions
-	k.Cdc.MustUnmarshalBinaryBare(store.Get(key), &reactions)
-
-	// Check if the user exists
-	if !reactions.ContainsReactionFrom(user, value) {
-		return fmt.Errorf("cannot remove the reaction with value %s from user %s as it does not exist",
-			value, user)
-	}
-
-	// Remove and save the reactions list
-	if newLikes, edited := reactions.RemoveReaction(user, value); edited {
-		if len(newLikes) == 0 {
-			store.Delete(key)
-		} else {
-			store.Set(key, k.Cdc.MustMarshalBinaryBare(&newLikes))
-		}
-	}
-
-	return nil
-}
-
-// GetPostReactions returns the list of reactions that has been associated to the post having the given id
-// nolint: interfacer
-func (k Keeper) GetPostReactions(ctx sdk.Context, postID types.PostID) types.Reactions {
-	store := ctx.KVStore(k.StoreKey)
-
-	var reactions types.Reactions
-	k.Cdc.MustUnmarshalBinaryBare(store.Get(types.PostReactionsStoreKey(postID)), &reactions)
-	return reactions
-}
-
-// GetReactions allows to returns the list of reactions that have been stored inside the given context
-func (k Keeper) GetReactions(ctx sdk.Context) map[types.PostID]types.Reactions {
-	store := ctx.KVStore(k.StoreKey)
-	iterator := sdk.KVStorePrefixIterator(store, types.PostReactionsStorePrefix)
-
-	reactionsData := map[types.PostID]types.Reactions{}
-	for ; iterator.Valid(); iterator.Next() {
-		var postLikes types.Reactions
-		k.Cdc.MustUnmarshalBinaryBare(iterator.Value(), &postLikes)
-		idBytes := bytes.TrimPrefix(iterator.Key(), types.PostReactionsStorePrefix)
-		postID, err := types.ParsePostID(string(idBytes))
-		if err != nil {
-			// This should never verify
-			panic(err)
-		}
-
-		reactionsData[postID] = postLikes
-	}
-
-	return reactionsData
 }
