+++ x/posts/internal/keeper/keeper_test.go
@@ -1,7 +1,6 @@
 package keeper_test
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
@@ -555,413 +554,6 @@ func TestKeeper_GetPostsFiltered(t *testing.T) {
 			for index, post := range result {
 				require.True(t, test.expected[index].Equals(post))
 			}
-		})
-	}
-}
-
-func TestKeeper_SavePollPostAnswers(t *testing.T) {
-	user, err := sdk.AccAddressFromBech32("cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns")
-	require.NoError(t, err)
-
-	user2, err := sdk.AccAddressFromBech32("cosmos1y54exmx84cqtasvjnskf9f63djuuj68p7hqf47")
-	require.NoError(t, err)
-
-	answers := []types.AnswerID{types.AnswerID(1), types.AnswerID(2)}
-	answers2 := []types.AnswerID{types.AnswerID(1)}
-
-	tests := []struct {
-		name               string
-		postID             types.PostID
-		userAnswersDetails types.UserAnswer
-		previousUsersAD    types.UserAnswers
-		expUsersAD         types.UserAnswers
-	}{
-		{
-			name:               "Save answers with no previous answers in this context",
-			postID:             types.PostID(1),
-			userAnswersDetails: types.NewUserAnswer(answers, user),
-			previousUsersAD:    nil,
-			expUsersAD:         types.UserAnswers{types.NewUserAnswer(answers, user)},
-		},
-		{
-			name:               "Save new answers",
-			postID:             types.PostID(1),
-			userAnswersDetails: types.NewUserAnswer(answers2, user2),
-			previousUsersAD:    types.UserAnswers{types.NewUserAnswer(answers, user)},
-			expUsersAD: types.UserAnswers{
-				types.NewUserAnswer(answers, user),
-				types.NewUserAnswer(answers2, user2),
-			},
-		},
-	}
-
-	for _, test := range tests {
-
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			ctx, k := SetupTestInput()
-			store := ctx.KVStore(k.StoreKey)
-
-			if test.previousUsersAD != nil {
-				store.Set(types.PollAnswersStoreKey(test.postID), k.Cdc.MustMarshalBinaryBare(test.previousUsersAD))
-			}
-
-			k.SavePollAnswers(ctx, test.postID, test.userAnswersDetails)
-
-			var actualUsersAnswersDetails types.UserAnswers
-			answersBz := store.Get(types.PollAnswersStoreKey(test.postID))
-			k.Cdc.MustUnmarshalBinaryBare(answersBz, &actualUsersAnswersDetails)
-			require.Equal(t, test.expUsersAD, actualUsersAnswersDetails)
-		})
-	}
-}
-
-func TestKeeper_GetPostPollAnswersDetails(t *testing.T) {
-	user, err := sdk.AccAddressFromBech32("cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns")
-	require.NoError(t, err)
-
-	answers := []types.AnswerID{types.AnswerID(1), types.AnswerID(2)}
-
-	tests := []struct {
-		name          string
-		postID        types.PostID
-		storedAnswers types.UserAnswers
-	}{
-		{
-			name:          "No answers returns empty list",
-			postID:        types.PostID(1),
-			storedAnswers: nil,
-		},
-		{
-			name:          "Answers returned correctly",
-			postID:        types.PostID(1),
-			storedAnswers: types.UserAnswers{types.NewUserAnswer(answers, user)},
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			ctx, k := SetupTestInput()
-
-			if test.storedAnswers != nil {
-				k.SavePollAnswers(ctx, test.postID, test.storedAnswers[0])
-			}
-
-			actualPostPollAnswers := k.GetPollAnswers(ctx, test.postID)
-
-			require.Equal(t, test.storedAnswers, actualPostPollAnswers)
-		})
-	}
-}
-
-func TestKeeper_GetPostPollAnswersByUser(t *testing.T) {
-	user, err := sdk.AccAddressFromBech32("cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns")
-	require.NoError(t, err)
-
-	user2, err := sdk.AccAddressFromBech32("cosmos1jlhazemxvu0zn9y77j6afwmpf60zveqw5480l2")
-	require.NoError(t, err)
-
-	answers := []types.AnswerID{types.AnswerID(1), types.AnswerID(2)}
-
-	tests := []struct {
-		name          string
-		storedAnswers types.UserAnswer
-		postID        types.PostID
-		user          sdk.AccAddress
-		expAnswers    []types.AnswerID
-	}{
-		{
-			name:          "No answers for user returns nil",
-			storedAnswers: types.NewUserAnswer(answers, user),
-			postID:        types.PostID(1),
-			user:          user2,
-			expAnswers:    nil,
-		},
-		{
-			name:          "Matching user returns answers made by him",
-			storedAnswers: types.NewUserAnswer(answers, user),
-			postID:        types.PostID(1),
-			user:          user,
-			expAnswers:    answers,
-		},
-	}
-
-	for _, test := range tests {
-		ctx, k := SetupTestInput()
-		k.SavePollAnswers(ctx, test.postID, test.storedAnswers)
-
-		actualPostPollAnswers := k.GetPollAnswersByUser(ctx, test.postID, test.user)
-		require.Equal(t, test.expAnswers, actualPostPollAnswers)
-	}
-}
-
-func TestKeeper_GetAnswersDetailsMap(t *testing.T) {
-	user, err := sdk.AccAddressFromBech32("cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4")
-	require.NoError(t, err)
-
-	user2, err := sdk.AccAddressFromBech32("cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae")
-	require.NoError(t, err)
-
-	answers := []types.AnswerID{types.AnswerID(1), types.AnswerID(2)}
-
-	tests := []struct {
-		name    string
-		usersAD map[types.PostID]types.UserAnswers
-	}{
-		{
-			name:    "Empty users answers details data are returned correctly",
-			usersAD: map[types.PostID]types.UserAnswers{},
-		},
-		{
-			name: "Non empty users answers details data are returned correcly",
-			usersAD: map[types.PostID]types.UserAnswers{
-				types.PostID(1): {
-					types.NewUserAnswer(answers, user),
-					types.NewUserAnswer(answers, user2),
-				},
-				types.PostID(2): {
-					types.NewUserAnswer(answers, user2),
-				},
-			},
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			ctx, k := SetupTestInput()
-			store := ctx.KVStore(k.StoreKey)
-			for postID, userAD := range test.usersAD {
-				store.Set(types.PollAnswersStoreKey(postID), k.Cdc.MustMarshalBinaryBare(userAD))
-			}
-
-			usersADData := k.GetPollAnswersMap(ctx)
-			require.Equal(t, test.usersAD, usersADData)
-		})
-	}
-}
-
-// -------------
-// --- Reactions
-// -------------
-
-func TestKeeper_SaveReaction(t *testing.T) {
-	liker, err := sdk.AccAddressFromBech32("cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4")
-	require.NoError(t, err)
-
-	otherLiker, err := sdk.AccAddressFromBech32("cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae")
-	require.NoError(t, err)
-
-	tests := []struct {
-		name           string
-		storedLikes    types.Reactions
-		postID         types.PostID
-		like           types.Reaction
-		error          error
-		expectedStored types.Reactions
-	}{
-		{
-			name:           "Reaction from same user already present returns expError",
-			storedLikes:    types.Reactions{types.NewReaction("like", liker)},
-			postID:         types.PostID(10),
-			like:           types.NewReaction("like", liker),
-			error:          fmt.Errorf("cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4 has already reacted with like to the post with id 10"),
-			expectedStored: types.Reactions{types.NewReaction("like", liker)},
-		},
-		{
-			name:           "First liker is stored properly",
-			storedLikes:    types.Reactions{},
-			postID:         types.PostID(15),
-			like:           types.NewReaction("like", liker),
-			error:          nil,
-			expectedStored: types.Reactions{types.NewReaction("like", liker)},
-		},
-		{
-			name:        "Second liker is stored properly",
-			storedLikes: types.Reactions{types.NewReaction("like", liker)},
-			postID:      types.PostID(87),
-			like:        types.NewReaction("like", otherLiker),
-			error:       nil,
-			expectedStored: types.Reactions{
-				types.NewReaction("like", liker),
-				types.NewReaction("like", otherLiker),
-			},
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			ctx, k := SetupTestInput()
-
-			store := ctx.KVStore(k.StoreKey)
-			if len(test.storedLikes) != 0 {
-				store.Set(types.PostReactionsStoreKey(test.postID), k.Cdc.MustMarshalBinaryBare(&test.storedLikes))
-			}
-
-			err := k.SaveReaction(ctx, test.postID, test.like)
-			require.Equal(t, test.error, err)
-
-			var stored types.Reactions
-			k.Cdc.MustUnmarshalBinaryBare(store.Get(types.PostReactionsStoreKey(test.postID)), &stored)
-			require.Equal(t, test.expectedStored, stored)
-		})
-	}
-}
-
-func TestKeeper_RemoveReaction(t *testing.T) {
-	liker, err := sdk.AccAddressFromBech32("cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4")
-	require.NoError(t, err)
-
-	tests := []struct {
-		name           string
-		storedLikes    types.Reactions
-		postID         types.PostID
-		liker          sdk.AccAddress
-		value          string
-		error          error
-		expectedStored types.Reactions
-	}{
-		{
-			name:           "Reaction from same liker is removed properly",
-			storedLikes:    types.Reactions{types.NewReaction("like", liker)},
-			postID:         types.PostID(10),
-			liker:          liker,
-			value:          "like",
-			error:          nil,
-			expectedStored: types.Reactions{},
-		},
-		{
-			name:           "Non existing reaction returns error - Creator",
-			storedLikes:    types.Reactions{},
-			postID:         types.PostID(15),
-			liker:          liker,
-			value:          "like",
-			error:          fmt.Errorf("cannot remove the reaction with value like from user cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4 as it does not exist"),
-			expectedStored: types.Reactions{},
-		},
-		{
-			name:           "Non existing reaction returns error - Value",
-			storedLikes:    types.Reactions{types.NewReaction("like", liker)},
-			postID:         types.PostID(15),
-			liker:          liker,
-			value:          "reaction",
-			error:          fmt.Errorf("cannot remove the reaction with value reaction from user cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4 as it does not exist"),
-			expectedStored: types.Reactions{types.NewReaction("like", liker)},
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			ctx, k := SetupTestInput()
-
-			store := ctx.KVStore(k.StoreKey)
-			if len(test.storedLikes) != 0 {
-				store.Set(types.PostReactionsStoreKey(test.postID), k.Cdc.MustMarshalBinaryBare(&test.storedLikes))
-			}
-
-			err := k.RemoveReaction(ctx, test.postID, test.liker, test.value)
-			require.Equal(t, test.error, err)
-
-			var stored types.Reactions
-			k.Cdc.MustUnmarshalBinaryBare(store.Get(types.PostReactionsStoreKey(test.postID)), &stored)
-
-			require.Len(t, stored, len(test.expectedStored))
-			for index, like := range test.expectedStored {
-				require.Equal(t, like, stored[index])
-			}
-		})
-	}
-}
-
-func TestKeeper_GetPostLikes(t *testing.T) {
-	liker, err := sdk.AccAddressFromBech32("cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4")
-	require.NoError(t, err)
-
-	otherLiker, err := sdk.AccAddressFromBech32("cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae")
-	require.NoError(t, err)
-
-	tests := []struct {
-		name   string
-		likes  types.Reactions
-		postID types.PostID
-	}{
-		{
-			name:   "Empty list are returned properly",
-			likes:  types.Reactions{},
-			postID: types.PostID(10),
-		},
-		{
-			name: "Valid list of likes is returned properly",
-			likes: types.Reactions{
-				types.NewReaction("like", otherLiker),
-				types.NewReaction("like", liker),
-			},
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			ctx, k := SetupTestInput()
-
-			for _, l := range test.likes {
-				err := k.SaveReaction(ctx, test.postID, l)
-				require.NoError(t, err)
-			}
-
-			stored := k.GetPostReactions(ctx, test.postID)
-
-			require.Len(t, stored, len(test.likes))
-			for _, l := range test.likes {
-				require.Contains(t, stored, l)
-			}
-		})
-	}
-}
-
-func TestKeeper_GetLikes(t *testing.T) {
-	liker1, err := sdk.AccAddressFromBech32("cosmos1s3nh6tafl4amaxkke9kdejhp09lk93g9ev39r4")
-	require.NoError(t, err)
-
-	liker2, err := sdk.AccAddressFromBech32("cosmos15lt0mflt6j9a9auj7yl3p20xec4xvljge0zhae")
-	require.NoError(t, err)
-
-	tests := []struct {
-		name  string
-		likes map[types.PostID]types.Reactions
-	}{
-		{
-			name:  "Empty likes data are returned correctly",
-			likes: map[types.PostID]types.Reactions{},
-		},
-		{
-			name: "Non empty likes data are returned correcly",
-			likes: map[types.PostID]types.Reactions{
-				types.PostID(5): {
-					types.NewReaction("like", liker1),
-					types.NewReaction("like", liker2),
-				},
-				types.PostID(10): {
-					types.NewReaction("like", liker1),
-				},
-			},
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			ctx, k := SetupTestInput()
-			store := ctx.KVStore(k.StoreKey)
-			for postID, likes := range test.likes {
-				store.Set(types.PostReactionsStoreKey(postID), k.Cdc.MustMarshalBinaryBare(likes))
-			}
-
-			likesData := k.GetReactions(ctx)
-			require.Equal(t, test.likes, likesData)
 		})
 	}
 }
