+++ x/posts/legacy/v0.1.0/types.go
@@ -27,7 +27,7 @@ type Post struct {
 	Owner             sdk.AccAddress `json:"owner"`              // Creator of the post
 }
 
-// Reaction is a struct of a user like
+// PostReaction is a struct of a user like
 type Like struct {
 	Created sdk.Int        `json:"created"` // Block height at which the like was created
 	Owner   sdk.AccAddress `json:"owner"`   // User that has inserted the like
