+++ x/posts/legacy/v0.3.0/types.go
@@ -39,8 +39,8 @@ type Post struct {
 	Creator        sdk.AccAddress `json:"creator"`                 // Creator of the Post
 }
 
-// Reaction is a struct of a user reaction to a post
+// PostReaction is a struct of a user reaction to a post
 type Reaction struct {
 	Owner sdk.AccAddress `json:"owner"` // User that has created the reaction
-	Value string         `json:"value"` // Reaction of the reaction
+	Value string         `json:"value"` // PostReaction of the reaction
 }
